add: test ligin api
